# swagger.yaml
openapi: "3.0.0"
info:
  title: learning express and swagger
  description: Becoming a backend pro developer
  version: 1.1.0
  contact:
    email: punit644@gmail.com
    url: "https://punit-bathija.netlify.app"

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for local host - secured
#   - url: "http://localhost:4000/api/v1"
#     description: for local host - unsecured

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for local host
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /app:
    get:
      tags:
        - String
      summary: returns greeting message from the admin
      responses:
        200:
          description: All went well Success!
          content:
            application/json:
              schema:
                type: string
                example: "my app"
        400:
          description: Bad Request
        500:
          description: Internal server error

  /appobject:
    get:
      tags:
        - Object
      summary: returns an unique course
      responses:
        200:
          description: All went well Success!
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number

        400:
          description: Bad Request
        500:
          description: Internal server error

  /courses:
    get:
      tags:
        - Array
      summary: returns all course's
      responses:
        200:
          description: All went well Success!
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number

        400:
          description: Bad Request
        500:
          description: Internal server error

  /mycourse/{courseId}:
    get:
      tags:
        - String
      summary: returns course matching with id sent in parameter
      parameters:
        - name: courseId
          in: path
          required: true
          default: 22
          schema:
            type: string
      responses:
        200:
          description: All went well Success!
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: number

        400:
          description: Bad Request
        500:
          description: Internal server error

  /addCourse:
    post:
      tags:
        - String
      summary: add's new course to your courses
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                price:
                  type: number
      responses:
        200:
          description: All went well Success!
          content:
            application/json:
              schema:
                type: boolean
        400:
          description: Bad Request
        500:
          description: Internal server error